requirements.txt
# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Databases
cassandra-driver==3.28.0
pymongo==4.6.1
clickhouse-connect==0.6.23
neo4j==5.15.0
redis==5.0.1

# Streaming
kafka-python==2.0.2

# Spark
pyspark==3.5.0

# Orchestration
apache-airflow==2.8.0
prefect==2.14.10

# Data Quality
great-expectations==0.18.8

# Visualization
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2

# ML & Analytics
scikit-learn==1.3.2
xgboost==2.0.3

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
kaggle==1.5.16

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
jupyter==1.0.0
